+++ src/pyramid_bimt/tests/test_AuditLogEventType.py
@@ -2,6 +2,8 @@
 """Tests for AuditLogEventType."""
 
 from pyramid import testing
+from pyramid_basemodel import Session
+from pyramid_bimt.testing import initTestingDB
 
 import unittest
 
@@ -10,8 +12,10 @@ class TestAuditLogEventType(unittest.TestCase):
 
     def setUp(self):
         self.config = testing.setUp()
+        initTestingDB()
 
     def tearDown(self):
+        Session.remove()
         testing.tearDown()
 
     def test_default_event_types(self):
@@ -42,3 +46,29 @@ class TestAuditLogEventType(unittest.TestCase):
         self.assertEqual(types[5].name, 'UserLoggedOut')
         self.assertEqual(types[5].title, 'User Logged Out')
         self.assertEqual(types[5].description, 'Emitted whenever a user logs out.')  # noqa
+
+    def test_get_event_type_by_id(self):
+        from pyramid_bimt.models import AuditLogEventType
+
+        self.assertEqual(AuditLogEventType.by_id(1).name, 'UserChangedPassword')  # noqa
+        self.assertEqual(AuditLogEventType.by_id(1).title, 'User Changed Password')  # noqa
+        self.assertEqual(AuditLogEventType.by_id(1).description, 'Emitted whenever a user changes its password.')  # noqa
+
+    def test_get_event_type_get_all(self):
+        from pyramid_bimt.scripts.populate import default_audit_log_event_types
+        from pyramid_bimt.models import AuditLogEventType
+        event_types = default_audit_log_event_types()
+        db_event_types = AuditLogEventType.get_all()
+        self.assertEqual(event_types[0].name, db_event_types[0].name)
+        self.assertEqual(db_event_types.count(), len(event_types))
+        self.assertEqual(
+            AuditLogEventType.get_all(filter_by={'name': 'UserLoggedOut'})[0].name,  # noqa
+            'UserLoggedOut',
+        )
+
+    def test_get_event_type_by_name(self):
+        from pyramid_bimt.models import AuditLogEventType
+
+        self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').id, 1)  # noqa
+        self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').title, 'User Changed Password')  # noqa
+        self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').description, 'Emitted whenever a user changes its password.')  # noqa