+++ src/pyramid_bimt/tests/test_views.py
@@ -18,6 +18,7 @@ class TestLoginViewsFunctional(unittest.TestCase):
         settings = {
             'bimt.app_title': 'BIMT',
             'bimt.disabled_user_redirect_path': '/settings',
+            'mail.default_sender': 'test_sender'
         }
         self.config = testing.setUp(settings=settings)
         initTestingDB()
@@ -39,6 +40,14 @@ class TestLoginViewsFunctional(unittest.TestCase):
         self.assertIn('302 Found', resp.text)
         resp = resp.follow()
         self.assertIn('Login successful.', resp.text)
+
+    def test_login_wrong_password(self):
+        resp = self.testapp.get('/login', status=200)
+        self.assertIn('<h1>Login</h1>', resp.text)
+        resp.form['email'] = 'ONE@bar.com'
+        resp.form['password'] = 'foo'
+        resp = resp.form.submit('login')
+        self.assertIn('Login failed.', resp.text)
 
     def test_login_disabled(self):
         from pyramid_bimt.models import User
@@ -62,6 +71,55 @@ class TestLoginViewsFunctional(unittest.TestCase):
                 'This test should fail with message /settings! '
                 'But it fails with message {}'.format(cm.exception.message)
             )
+
+    class _MessageMatcher(object):
+        def __init__(self, compare, msg):
+            self.msg = msg
+            self.compare = compare
+
+        def __eq__(self, other):
+            return self.compare(self.msg, other)
+
+    @mock.patch('pyramid_bimt.views.auth.get_mailer')
+    def test_login_reset_password(self, get_mailer):
+        resp = self.testapp.get('/login', status=200)
+        from pyramid_mailer.message import Message
+        self.assertIn('<h1>Login</h1>', resp.text)
+        resp.form['email'] = 'ONE@bar.com'
+        resp = resp.form.submit('reset_password')
+        self.assertIn('302 Found', resp.text)
+        resp = resp.follow()
+        self.assertIn('A new password was sent to your email.', resp.text)
+        message = Message(
+            subject='{} Password Reset'.format('BIMT'),
+            sender='test_sender',
+            recipients=['one@bar.com', ],
+            body='test',
+        )
+        match_message = self._MessageMatcher(compare_message, message)
+        get_mailer().send.assert_called_with(match_message)
+
+        with self.assertRaises(AssertionError):
+            message.subject = 'test'
+            match_message = self._MessageMatcher(compare_message, message)
+            get_mailer().send.assert_called_with(match_message)
+
+    def test_login_reset_password_no_user(self):
+        resp = self.testapp.get('/login', status=200)
+        self.assertIn('<h1>Login</h1>', resp.text)
+        resp.form['email'] = 'foo@bar.com'
+        resp = resp.form.submit('reset_password')
+        self.assertIn('Password reset failed. Make sure you have correctly '
+                      'entered your email address.', resp.text)
+
+
+def compare_message(self, other):
+    if (self.subject == other.subject and
+            self.sender == other.sender and
+            self.recipients == other.recipients):
+        return True
+    else:
+        return False
 
 
 class TestUserView(unittest.TestCase):
@@ -114,7 +172,12 @@ class TestUserViewFunctional(unittest.TestCase):
     def test_disable_enable_user(self):
         self.config.testing_securitypolicy(
             userid='admin@bar.com', permissive=True)
+        # Ran 2 times, so we check what happens if already disabled
         self.testapp.get('/users/1/disable', status=302)
+        self.testapp.get('/users/1/disable', status=302)
+
+        # Ran 2 times, so we check what happens if already disabled
+        self.testapp.get('/users/1/enable', status=302)
         self.testapp.get('/users/1/enable', status=302)
 
     def test_user_actions(self):
@@ -182,11 +245,128 @@ class TestEditUserViewFunctional(unittest.TestCase):
         self.assertIn('<td>two@bar.com</td>', resp.text)
         self.assertIn('<td>One Two</td>', resp.text)
 
+    def test_edit_user_no_group(self):
+        self.config.testing_securitypolicy(
+            userid='one@bar.com', permissive=True)
+        resp = self.testapp.get('/users/1/edit', status=200)
+        self.assertIn('<h1>Edit User</h1>', resp.text)
+
+        form = resp.forms['useredit']
+        form['fullname'] = u'One Two'
+        form['email'] = u'TWO@bar.com'
+        form.set('Group', u'not a group', index=0)
+        resp = form.submit('save')
+
     def test_edit_no_user(self):
         self.config.testing_securitypolicy(
             userid='one@bar.com', permissive=True)
         self.assertRaises(
             HTTPNotFound, self.testapp.get, '/users/123456789/edit')
+
+
+class TestAuditLogView(unittest.TestCase):
+    def setUp(self):
+        settings = {
+            'bimt.app_title': 'BIMT',
+        }
+        self.config = testing.setUp(settings=settings)
+        initTestingDB()
+        configure(self.config)
+        self.request = testing.DummyRequest()
+        app = self.config.make_wsgi_app()
+        self.testapp = webtest.TestApp(app)
+
+    def tearDown(self):
+        Session.remove()
+        testing.tearDown()
+
+    def test_audit_log(self):
+        self.config.testing_securitypolicy(
+            userid='one@bar.com', permissive=True)
+        resp = self.testapp.get('/audit-log', status=200)
+        self.assertIn('<h1>Audit Log</h1>', resp.text)
+
+    def test_audit_log_delete(self):
+        from pyramid_bimt.models import AuditLogEventType
+        from pyramid_bimt.models import AuditLogEntry
+        from pyramid_bimt.views.auditlog import audit_log_delete
+        import transaction
+        self.config.testing_securitypolicy(
+            userid='one@bar.com', permissive=True)
+        request = self.request
+        entry = AuditLogEntry(
+            user_id=1,
+            event_type_id=AuditLogEventType.by_name('UserCreated').id,
+        )
+        Session.add(entry)
+        request.context = entry
+        transaction.commit()
+        resp = audit_log_delete(request)
+        self.assertIn('/audit-log', resp.location)
+
+    def test_audit_log_add(self):
+        self.config.testing_securitypolicy(
+            userid='one@bar.com', permissive=True)
+        resp = self.testapp.get('/audit-log/add', status=200)
+        self.assertIn('<h1>Add Audit log entry</h1>', resp.text)
+
+    def test_audit_log_add_submit_success(self):
+        from pyramid_bimt.views.auditlog import AuditLogAddEntryForm
+        import datetime
+        audit_log_add_view = AuditLogAddEntryForm(self.request)
+        form_values = {
+            'timestamp': datetime.datetime.now(),
+            'user_id': 1,
+            'event_type_id': 1,
+            'comment': 'testing'
+        }
+        resp = audit_log_add_view.submit_success(form_values)
+        self.assertIn('/audit-log', resp.location)
+
+
+class TestLogoutView(unittest.TestCase):
+    def setUp(self):
+        settings = {
+            'bimt.app_title': 'BIMT',
+        }
+        self.config = testing.setUp(settings=settings)
+        initTestingDB()
+        configure(self.config)
+
+    def tearDown(self):
+        Session.remove()
+        testing.tearDown()
+
+    def test_logout_view(self):
+        from pyramid_bimt.models import User
+        from pyramid_bimt.views.auth import logout
+        context = User.by_email('admin@bar.com')
+        request = testing.DummyRequest(layout_manager=mock.Mock())
+        request.user = context
+        self.assertEqual(
+            logout(context, request).location,
+            request.params.get('came_from', request.application_url)
+        )
+
+
+class TestForbiddenRedirect(unittest.TestCase):
+    def setUp(self):
+        settings = {
+            'bimt.app_title': 'BIMT',
+        }
+        self.config = testing.setUp(settings=settings)
+        configure(self.config)
+
+    def tearDown(self):
+        testing.tearDown()
+
+    def test_forbidden_redirect_view(self):
+        from pyramid_bimt.views.auth import forbidden_redirect
+        request = testing.DummyRequest(layout_manager=mock.Mock())
+        self.assertEqual(
+            forbidden_redirect(None, request).location,
+            request.route_url('login', _query={'came_from': request.url}),
+        )
 
 
 class TestConfig(unittest.TestCase):